Update MetrolookTemplate.php
This fixes the javascript problem also add logo back but css is needed to enable it. will add the css and fix a issue where if you add the css it wont show properly if people would like the css please add your support for it to issue but please know if enable the logo will be like vector where it has a gray background and is not on the top bar but on the bottom. this also fixes other issues.