Update patch set 2
Patch Set 2:

So if you look at the console on https://integration.wikimedia.org/ci/job/php-composer-test/7938/console  the sequence of commands is roughly:

 composer update --no-progress --prefer-dist --profile
 composer test

And that manages to run composer.


On Jenkins, the composer test command is not run for you.  I have proposed to allow the job to be run for your account with https://gerrit.wikimedia.org/r/#/c/222187/

Patch-set: 2
Label: Code-Review=0