Update patch set 2
Patch Set 2: Code-Review-1

It probably works as intended as far as desktop is concerned, but it seems a bit choppy on mobile. All is fine and dandy at least as long as the hamburger (sidebar) menu isn't open. When it is, it seems that its z-index is not high enough (well, actually it doesn't have a z-index set IIRC), so the search bar overlaps it. This is easy enough to fix: div#mw-panel { z-index: 1; }
In fact, my Labs wiki at http://social-tools.wmflabs.org/w/index.php?title=Main_Page&useskin=metrolook which is currently running this patch also has that small z-index fix.

The choppiness comes in when you scroll the page on a mobile device with the hamburger menu open. When you initially start scrolling down, there is a strange, visible gap in the sidebar area.

And finally some more z-index woes...I would imagine that the content action links ("View source", "History", etc.) would have a higher z-index than the sidebar menu (#mw-panel). If I recall correctly, that is indeed the case...yet on my Labs site the sidebar overlaps the content action tools (provided that you've toggled the content action tools to be active, that is).

For reference:
* http://social-tools.wmflabs.org/w/index.php?title=Main_Page&useskin=metrolook -- my Labs wiki, running master version of Metrolook with this patch applied
* http://mwmsg.shoutwiki.com/w/index.php?title=Main_Page&useskin=metrolook -- ShoutWiki's Interface Messages Wiki, running the 17 July 2015 version of Metrolook (4298ae7785ac3cfc2511f29782fed21ddf2440d1)

My test device was a Nokia X, running Nokia's fork of Android 4.1.2 with Firefox for Android v. 39.0 as the browser.

Patch-set: 2