Replace deprecated `background-image-svg()` mixin calls
Use normal `background-image` properties with SVGs now
that IE 8 and Android 2.1 are removed from Grade C.
This patch is only caring about this specific mixin and its image
dependencies. There are a number of IE 8 and even IE 6/7 hacks in
Metrolook as is. An alternative would be to put the same mixin into
the theme and keep it as is, but I don't think it's worth it to
continue support for the 11 year old browser on master.
Also removing PNG fallback images, exception is 'user-icon-20.png',
which is used in a specific fallback that would need to be removed in
a separate patch.

Bumping MW version requirements and skin version in this process.

Bug: T248062
Change-Id: I0a1c8a3717cace081d1227317154966c8f0b2760